Rename metadata use cases and move it to domain
