+++ src/pages/sync-rules-list/SyncRulesListPage.tsx
@@ -25,7 +25,7 @@ import { TestWrapper } from "../../components/test-wrapper/TestWrapper";
 import { AggregatedSync } from "../../logic/sync/aggregated";
 import { EventsSync } from "../../logic/sync/events";
 import { SyncronizationClass } from "../../logic/sync/generic";
-import { MetadataSync } from "../../logic/sync/metadata";
+import { MetadataSyncUseCase } from "../../data/metadata/usecases/MetadataSyncUseCase";
 import Instance from "../../models/instance";
 import SyncReport from "../../models/syncReport";
 import SyncRule from "../../models/syncRule";
@@ -51,7 +51,7 @@ const config: {
 } = {
     metadata: {
         title: i18n.t("Metadata Synchronization Rules"),
-        SyncClass: MetadataSync,
+        SyncClass: MetadataSyncUseCase,
     },
     aggregated: {
         title: i18n.t("Aggregated Data Synchronization Rules"),
